pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastdlp"
version = "0.1.0"
description = "Fast Data Loading Package for Robot Trajectories"
authors = [
    { name = "Max Fu", email = "max.fu.letian@berkeley.edu" }
]
dependencies = [
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "torchvision>=0.10.0",
    "zarr>=2.10.0",
    "Pillow>=8.0.0",
    "transformers>=4.15.0",
    "tqdm>=4.62.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "isort>=5.0",
    "mypy>=0.910",
]